backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
supabase==2.3.0
python-dotenv==1.0.0
